# docs/spec/openapi.yml
openapi: 3.0.2
servers:
  - url: /v1
info:
  version: 1.0.0
  title: Percona Everest schema
  description: |
    # Authentication
    All requests to Everest API require `Authorization: Bearer <token>` header with a valid token in plain-text.
    
    The token can be obtained by using `everestctl token reset` which resets the token and prints it to the screen.
...
paths:
...
  '/resources':
    get:
      tags:
        - k8s
      summary: Get the capacity and available resources of a kubernetes cluster
      description: Get the capacity and available resources of a kubernetes cluster
      operationId: getKubernetesClusterResources
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KubernetesClusterResources'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  '/namespaces/{namespace}/database-clusters/{name}':
    get:
      tags:
        - databaseCluster
      summary: Get the specified database cluster
      description: Get the specified database cluster
      operationId: getDatabaseCluster
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCluster'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - databaseCluster
      summary: Replace the specified database cluster
      description: Replace the specified database cluster
      operationId: updateDatabaseCluster
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCluster'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        description: The database cluster object to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseCluster'
    delete:
      tags:
        - databaseCluster
      summary: Delete the specified database cluster
      description: Delete the specified database cluster
      operationId: deleteDatabaseCluster
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/io.k8s.apimachinery.pkg.apis.meta.v1.Status_v2'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  '/namespaces/{namespace}/database-clusters/{name}/restores':
    get:
      tags:
        - databaseClusterRestore
      summary: List of the created database cluster restores
      description: List of the created database cluster restores
      operationId: listDatabaseClusterRestores
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster. Can be found under Metadata["name"] of the DatabaseCluster object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestoreList'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  '/namespaces/{namespace}/database-cluster-restores':
    post:
      tags:
        - databaseClusterRestore
      summary: Create a database cluster restore
      description: Create a database cluster restore
      operationId: createDatabaseClusterRestore
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestore'
        '201':
          description: Created success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestore'
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestore'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        description: The database cluster restore object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseClusterRestore'
  '/namespaces/{namespace}/database-cluster-restores/{name}':
    get:
      tags:
        - databaseClusterRestore
      summary: Returns the specified cluster restore
      description: Returns the specified cluster restore
      operationId: getDatabaseClusterRestore
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster restore. Can be found under Metadata["name"] of the DatabaseClusterRestore object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestore'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - databaseClusterRestore
      summary: Replace the specified cluster restore
      description: Replace the specified cluster restore
      operationId: updateDatabaseClusterRestore
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster restore. Can be found under Metadata["name"] of the DatabaseClusterRestore object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestore'
        '201':
          description: Created successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseClusterRestore'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        description: The database cluster restore object to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseClusterRestore'
    delete:
      tags:
        - databaseClusterRestore
      summary: Delete the specified cluster restore
      description: Delete the specified cluster restore
      operationId: deleteDatabaseClusterRestore
      parameters:
        - name: namespace
          in: path
          description: Name of the namespace
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name of the database cluster restore. Can be found under Metadata["name"] of the DatabaseClusterRestore object.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/io.k8s.apimachinery.pkg.apis.meta.v1.Status_v2'
        '400':
          description: Unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      description: Error response
      properties:
        message:
          type: string
...
    KubernetesClusterResources:
      type: object
      description: kubernetes cluster resources
      properties:
        capacity:
          type: object
          x-go-type-name: ResourcesCapacity
          properties:
            cpuMillis:
              type: number
              x-go-type: uint64
            memoryBytes:
              x-go-type: uint64
              type: number
            diskSize:
              type: number
              x-go-type: uint64
        available:
          type: object
          x-go-type-name: ResourcesAvailable
          properties:
            cpuMillis:
              type: number
              x-go-type: uint64
            memoryBytes:
              type: number
              x-go-type: uint64
            diskSize:
              type: number
              x-go-type: uint64
      required:
        - capacity
        - available
...
    DatabaseClusterRestoreList:
      description: DatabaseClusterRestoreList is an object that contains the list of the existing database cluster restores.
      properties:
        apiVersion:
          description: 'APIVersion defines the versioned schema of this representation
                  of an object. Servers should convert recognized schemas to the latest
                  internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
          type: string
        kind:
          description: 'Kind is a string value representing the REST resource this
                  object represents. Servers may infer this from the endpoint the client
                  submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/DatabaseClusterRestore'
        metadata:
          type: object
      type: object
...
    DatabaseClusterRestore:
      description: DatabaseClusterRestore is the Schema for the databaseclusterrestores API.
      properties:
        apiVersion:
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
          type: string
        kind:
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          type: string
        metadata:
          type: object
        spec:
          description: DatabaseClusterRestoreSpec defines the desired state of DatabaseClusterRestore.
          properties:
            dataSource:
              description: DataSource defines a data source for restoration.
              properties:
                backupSource:
                  description: BackupSource is the backup source to restore from
                  properties:
                    backupStorageName:
                      description: BackupStorageName is the name of the BackupStorage used for backups.
                      type: string
                    path:
                      description: Path is the path to the backup file/directory.
                      type: string
                  required:
                    - backupStorageName
                    - path
                  type: object
                dbClusterBackupName:
                  description: DBClusterBackupName is the name of the DB cluster backup to restore from
                  type: string
                pitr:
                  description: PITR is the point-in-time recovery configuration
                  properties:
                    date:
                      description: 'Date is the UTC date to recover to. The accepted format: "2006-01-02T15:04:05Z".'
                      type: string
                    type:
                      default: date
                      description: Type is the type of recovery.
                      enum:
                        - date
                        - latest
                      type: string
                  type: object
              type: object
            dbClusterName:
              description: DBClusterName defines the cluster name to restore.
              type: string
          required:
            - dataSource
            - dbClusterName
          type: object
        status:
          description: DatabaseClusterRestoreStatus defines the observed state of DatabaseClusterRestore.
          properties:
            completed:
              format: date-time
              type: string
            message:
              type: string
            state:
              description: RestoreState represents state of restoration.
              type: string
          type: object
      type: object

    # -------------------------
    # io.k8s.apimachinery.pkg.apis.meta.v1.Status_v2 spec is imported from k8s API,
    # please do not edit the definition below manually
    # -------------------------
    io.k8s.apimachinery.pkg.apis.meta.v1.Status_v2:
      description: Status is a return value for calls that don't return other objects.
      type: object
      properties:
        apiVersion:
          description: 'APIVersion defines the versioned schema of this representation
            of an object. Servers should convert recognized schemas to the latest internal
            value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
          type: string
        code:
          description: Suggested HTTP return code for this status, 0 if not set.
          type: integer
          format: int32
        details:
          description: Extended data associated with the reason.  Each reason may define
            its own extended details. This field is optional and the data returned is
            not guaranteed to conform to any schema except that defined by the reason
            type.
          "$ref": "#/components/schemas/io.k8s.apimachinery.pkg.apis.meta.v1.StatusDetails_v2"
        kind:
          description: 'Kind is a string value representing the REST resource this object
            represents. Servers may infer this from the endpoint the client submits requests
            to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          type: string
        message:
          description: A human-readable description of the status of this operation.
          type: string
        metadata:
          description: 'Standard list metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          "$ref": "#/components/schemas/io.k8s.apimachinery.pkg.apis.meta.v1.ListMeta"
        reason:
          description: A machine-readable description of why this operation is in the
            "Failure" status. If this value is empty there is no information available.
            A Reason clarifies an HTTP status code but does not override it.
          type: string
        status:
          description: 'Status of the operation. One of: "Success" or "Failure". More
            info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status'
          type: string

    # -------------------------
    # io.k8s.apimachinery.pkg.apis.meta.v1.ListMeta spec is imported from k8s API,
    # please do not edit the definition below manually
    # -------------------------
    io.k8s.apimachinery.pkg.apis.meta.v1.ListMeta:
      description: ListMeta describes metadata that synthetic resources must have, including
        lists and various status objects. A resource may have only one of {ObjectMeta,
        ListMeta}.
      type: object
      properties:
        continue:
          description: continue may be set if the user set a limit on the number of items
            returned, and indicates that the server has more data available. The value
            is opaque and may be used to issue another request to the endpoint that served
            this list to retrieve the next set of available objects. Continuing a consistent
            list may not be possible if the server configuration has changed or more than
            a few minutes have passed. The resourceVersion field returned when using this
            continue value will be identical to the value in the first response, unless
            you have received this token from an error message.
          type: string
        remainingItemCount:
          description: remainingItemCount is the number of subsequent items in the list
            which are not included in this list response. If the list request contained
            label or field selectors, then the number of remaining items is unknown and
            the field will be left unset and omitted during serialization. If the list
            is complete (either because it is not chunking or because this is the last
            chunk), then there are no more remaining items and this field will be left
            unset and omitted during serialization. Servers older than v1.15 do not set
            this field. The intended use of the remainingItemCount is *estimating* the
            size of a collection. Clients should not rely on the remainingItemCount to
            be set or to be exact.
          type: integer
          format: int64
        resourceVersion:
          description: 'String that identifies the server''s internal version of this
            object that can be used by clients to determine when objects have changed.
            Value must be treated as opaque by clients and passed unmodified back to the
            server. Populated by the system. Read-only. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency'
          type: string
        selfLink:
          description: |-
            selfLink is a URL representing this object. Populated by the system. Read-only.

            DEPRECATED Kubernetes will stop propagating this field in 1.20 release and the field is planned to be removed in 1.21 release.
          type: string

    # -------------------------
    # io.k8s.apimachinery.pkg.apis.meta.v1.StatusDetails_v2 spec is imported from k8s API,
    # please do not edit the definition below manually
    # -------------------------
    io.k8s.apimachinery.pkg.apis.meta.v1.StatusDetails_v2:
      description: StatusDetails is a set of additional properties that MAY be set by
        the server to provide additional information about a response. The Reason field
        of a Status object defines what attributes will be set. Clients must ignore fields
        that do not match the defined type of each attribute, and should assume that any
        attribute may be empty, invalid, or under defined.
      type: object
      properties:
        causes:
          description: The Causes array includes more details associated with the StatusReason
            failure. Not all StatusReasons may provide detailed causes.
          type: array
          items:
            "$ref": "#/components/schemas/io.k8s.apimachinery.pkg.apis.meta.v1.StatusCause"
        group:
          description: The group attribute of the resource associated with the status
            StatusReason.
          type: string
        kind:
          description: 'The kind attribute of the resource associated with the status
            StatusReason. On some operations may differ from the requested resource Kind.
            More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
          type: string
        name:
          description: The name attribute of the resource associated with the status StatusReason
            (when there is a single name which can be described).
          type: string
        retryAfterSeconds:
          description: If specified, the time in seconds before the operation should be
            retried. Some errors may indicate the client must take an alternate action
            - for those errors this field may indicate how long to wait before taking
            the alternate action.
          type: integer
          format: int32
        uid:
          description: 'UID of the resource. (when there is a single resource which can
            be described). More info: http://kubernetes.io/docs/user-guide/identifiers#uids'
          type: string

    # -------------------------
    # io.k8s.apimachinery.pkg.apis.meta.v1.StatusCause spec is imported from k8s API,
    # please do not edit the definition below manually
    # -------------------------
    io.k8s.apimachinery.pkg.apis.meta.v1.StatusCause:
      description: StatusCause provides more information about an api.Status failure,
        including cases when multiple errors are encountered.
      type: object
      properties:
        field:
          description: |-
            The field of the resource that has caused this error, as named by its JSON serialization. May include dot and postfix notation for nested attributes. Arrays are zero-indexed.  Fields may appear more than once in an array of causes due to fields having multiple errors. Optional.

            Examples:
              "name" - the field "name" on the current resource
              "items[0].name" - the field "name" on the first array entry in "items"
          type: string
        message:
          description: A human-readable description of the cause of the error.  This field
            may be presented as-is to a reader.
          type: string
        reason:
          description: A machine-readable description of the cause of the error. If this
            value is empty there is no information available.
          type: string